- adding a sleep to wait for the mysql container to be ready

+++ Makefile
@@ -3,12 +3,15 @@ composer:
       --volume $(shell pwd):/app \
       composer/composer install
 setup:
-	cp .env.example .env
+	if [ ! -f .env ]; then cp .env.example .env; fi
 	make composer
 	make start
+	make mysql-ready
 	./vendor/bin/sail artisan migrate
 	./vendor/bin/sail artisan key:generate
 	./vendor/bin/sail npm install
 	./vendor/bin/sail npx mix
 start:
 	./vendor/bin/sail up -d
+mysql-ready:
+	./vendor/bin/sail exec -T mysql bash -c 'until mysqladmin ping -h localhost --silent; do echo "Waiting for MySQL..."; sleep 1; done'
